# openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/api/v1/auth/login": {
      "post": {
        "description": "Public. Opens a session for an existing account. The access token comes back in the body; the refresh token is set only as an `httpOnly`, path-scoped cookie and is never returned in a body. A wrong password and an unknown address answer the same 401, so the route cannot be used to probe which addresses have accounts.",
        "operationId": "AuthController_login_v1",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LoginDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Session opened. Refresh cookie set on the response.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AccessTokenDto"
                }
              }
            }
          },
          "400": {
            "description": "`VALIDATION_ERROR` — Malformed body.",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    {
                      "$ref": "#/components/schemas/ApiErrorDto"
                    },
                    {
                      "properties": {
                        "code": {
                          "type": "string",
                          "enum": [
                            "VALIDATION_ERROR"
                          ]
                        }
                      }
                    }
                  ]
                }
              }
            }
          },
          "401": {
            "description": "`UNAUTHORIZED` — Wrong password, unknown address, or a disabled account.",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    {
                      "$ref": "#/components/schemas/ApiErrorDto"
                    },
                    {
                      "properties": {
                        "code": {
                          "type": "string",
                          "enum": [
                            "UNAUTHORIZED"
                          ]
                        }
                      }
                    }
                  ]
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Log in with email and password",
        "tags": [
          "auth"
        ]
      }
    },
...
    "/api/v1/auth/face/identities": {
      "post": {
        "description": "Registers a face token against the calling account so it can be used to log in later. Enrollment replaces the previous active identity rather than adding a second one; the replaced record is kept as history, so enrollments stay auditable.",
        "operationId": "FaceIdentityController_register_v1",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/FaceTokenDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Identity enrolled and now the active one.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/FaceIdentityDto"
                }
              }
            }
          },
          "400": {
            "description": "`VALIDATION_ERROR` — Malformed body.",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    {
                      "$ref": "#/components/schemas/ApiErrorDto"
                    },
                    {
                      "properties": {
                        "code": {
                          "type": "string",
                          "enum": [
                            "VALIDATION_ERROR"
                          ]
                        }
                      }
                    }
                  ]
                }
              }
            }
          },
          "401": {
            "description": "`UNAUTHORIZED` — Missing or invalid bearer token, or the account cannot be used for face login.",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    {
                      "$ref": "#/components/schemas/ApiErrorDto"
                    },
                    {
                      "properties": {
                        "code": {
                          "type": "string",
                          "enum": [
                            "UNAUTHORIZED"
                          ]
                        }
                      }
                    }
                  ]
                }
              }
            }
          },
          "403": {
            "description": "`FORBIDDEN` — Caller has not completed their profile.",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    {
                      "$ref": "#/components/schemas/ApiErrorDto"
                    },
                    {
                      "properties": {
                        "code": {
                          "type": "string",
                          "enum": [
                            "FORBIDDEN"
                          ]
                        }
                      }
                    }
                  ]
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Enroll a face identity",
        "tags": [
          "auth"
        ]
      }
    },
    "/api/v1/auth/face/login": {
      "post": {
        "description": "Public — the face token is the credential, so no password is involved. Opens a session: the access token comes back in the body, the refresh token only as an `httpOnly` path-scoped cookie. An unrecognized face and a disabled account answer the same 401, on purpose.",
        "operationId": "FaceIdentityController_login_v1",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/FaceTokenDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Session opened. Refresh token set as a cookie, never in the body.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AccessTokenDto"
                }
              }
            }
          },
          "400": {
            "description": "`VALIDATION_ERROR` — Malformed body.",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    {
                      "$ref": "#/components/schemas/ApiErrorDto"
                    },
                    {
                      "properties": {
                        "code": {
                          "type": "string",
                          "enum": [
                            "VALIDATION_ERROR"
                          ]
                        }
                      }
                    }
                  ]
                }
              }
            }
          },
          "401": {
            "description": "`UNAUTHORIZED` — Face not recognized, or the account cannot be used.",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    {
                      "$ref": "#/components/schemas/ApiErrorDto"
                    },
                    {
                      "properties": {
                        "code": {
                          "type": "string",
                          "enum": [
                            "UNAUTHORIZED"
                          ]
                        }
                      }
                    }
                  ]
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Log in with a face identity",
        "tags": [
          "auth"
        ]
      }
    },
...
  },
  "info": {
    "title": "Tu Seguridad API",
    "description": "Space-scoped video surveillance API. Every route outside `auth` and `health` resolves resources inside the caller space, taken from the bearer token, so an id from another space answers 404 rather than a result. Writes that change what the system watches are admin-only. Failures answer `{ statusCode, code, message }` — branch on `code`, not on the message.",
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      }
    },
    "schemas": {
      "ApiErrorDto": {
        "type": "object",
        "properties": {
          "statusCode": {
            "type": "number",
            "example": 404,
            "description": "HTTP status, the same value as the response status line."
          },
          "code": {
            "type": "string",
            "enum": [
              "VALIDATION_ERROR",
              "INVALID_ZONE",
              "UNAUTHORIZED",
              "FORBIDDEN",
              "NOT_FOUND",
              "CONFLICT",
              "UPSTREAM_ERROR",
              "UPSTREAM_TIMEOUT",
              "INTERNAL_ERROR"
            ],
            "example": "NOT_FOUND",
            "description": "Stable machine-readable code. Branch on this, not on the message."
          },
          "message": {
            "type": "string",
            "example": "Camera 7f3a not found",
            "description": "Human-readable reason. Never carries credentials, tokens or snapshot URLs."
          }
        },
        "required": [
          "statusCode",
          "code",
          "message"
        ]
      },
      "LoginDto": {
        "type": "object",
        "properties": {
          "email": {
            "type": "string",
            "example": "admin@example.com"
          },
          "password": {
            "type": "string",
            "example": "change-me"
          }
        },
        "required": [
          "email",
          "password"
        ]
      },
      "AccessTokenDto": {
        "type": "object",
        "properties": {
          "accessToken": {
            "type": "string",
            "description": "Bearer token for the `Authorization` header. Short-lived — rotate through `POST /auth/refresh` when it expires.",
            "example": "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJzdWIiOiI..."
          }
        },
        "required": [
          "accessToken"
        ]
      },
...
      "FaceTokenDto": {
        "type": "object",
        "properties": {
          "faceToken": {
            "type": "string",
            "description": "Opaque Face Auth identifier"
          }
        },
        "required": [
          "faceToken"
        ]
      },
      "FaceIdentityDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "createdAt": {
            "format": "date-time",
            "type": "string"
          }
        },
        "required": [
          "id",
          "createdAt"
        ]
      },
...
    }
  }
}